fix: refactor the stamps features
