+++ src/index.ts
@@ -2,7 +2,7 @@
 import { Application } from 'express'
 
 import { createApp } from './server'
-import { StampsManager } from './stamps'
+import { AutoBuyStampsManager, ExtendsStampManager, HardcodedStampsManager } from './stamps'
 import { getAppConfig, getServerConfig, getStampsConfig, EnvironmentVariables, getContentConfig } from './config'
 import { logger, subscribeLogServerRequests } from './logger'
 import { ContentManager } from './content'
@@ -28,9 +28,25 @@ async function main() {
 
   if (stampsConfig) {
     logger.debug('stamps config', stampsConfig)
-    const stampManager = new StampsManager()
-    logger.info('starting postage stamp manager')
-    stampManager.start(stampsConfig)
+    let stampManager: HardcodedStampsManager | AutoBuyStampsManager | ExtendsStampManager
+    const { mode } = stampsConfig
+
+    if (mode === 'hardcoded') {
+      stampManager = new HardcodedStampsManager()
+
+      logger.info('starting hardcoded postage stamp manager')
+      stampManager.start(stampsConfig)
+    } else if (mode === 'autobuy') {
+      stampManager = new AutoBuyStampsManager()
+
+      logger.info('starting autobuy postage stamp manager')
+      stampManager.start(stampsConfig)
+    } else {
+      stampManager = new ExtendsStampManager()
+
+      logger.info('starting extends postage stamp manager')
+      stampManager.start(stampsConfig)
+    }
     logger.info('starting the proxy')
     app = createApp(appConfig, stampManager)
   } else {
