+++ src/readiness.ts
@@ -1,7 +1,7 @@
 import { Bee, BeeDebug, Utils } from '@ethersphere/bee-js'
 import { ERROR_NO_STAMP, READINESS_TIMEOUT_MS } from './config'
 import { logger } from './logger'
-import { StampsManager } from './stamps'
+import { AutoBuyStampsManager, ExtendsStampManager, HardcodedStampsManager } from './stamps'
 import { getErrorMessage } from './utils'
 
 const MAX_CHUNK_SIZE = 4096
@@ -16,7 +16,7 @@ export enum ReadinessStatus {
 export async function checkReadiness(
   bee: Bee,
   beeDebug: BeeDebug,
-  stampManager?: StampsManager,
+  stampManager?: HardcodedStampsManager | AutoBuyStampsManager | ExtendsStampManager,
 ): Promise<ReadinessStatus> {
   if (stampManager) {
     const ready = await tryUploadingSingleChunk(bee, stampManager)
@@ -34,7 +34,10 @@ export async function checkReadiness(
   }
 }
 
-async function tryUploadingSingleChunk(bee: Bee, stampsManager: StampsManager): Promise<ReadinessStatus> {
+async function tryUploadingSingleChunk(
+  bee: Bee,
+  stampsManager: HardcodedStampsManager | AutoBuyStampsManager | ExtendsStampManager,
+): Promise<ReadinessStatus> {
   const chunk = makeChunk()
   try {
     await bee.uploadChunk(stampsManager.postageStamp, chunk, { timeout: READINESS_TIMEOUT_MS, deferred: true })
