+++ src/stamps/index.ts
@@ -1,0 +1,7 @@
+import { HardcodedStampsManager } from './hardcoded'
+import { AutoBuyStampsManager } from './autobuy'
+import { ExtendsStampManager } from './extends'
+
+export { HardcodedStampsManager }
+export { AutoBuyStampsManager }
+export { ExtendsStampManager }
