+++ test/readiness.spec.ts
@@ -1,6 +1,6 @@
 import request from 'supertest'
 import { createApp } from '../src/server'
-import { StampsManager } from '../src/stamps'
+import { HardcodedStampsManager } from '../src/stamps'
 
 describe('readiness', () => {
   test('should be ready without stamp management', async () => {
@@ -13,7 +13,7 @@ describe('readiness', () => {
   })
 
   test('should be ready with stamp management', async () => {
-    const stampManager = new StampsManager()
+    const stampManager = new HardcodedStampsManager()
     await stampManager.start({ mode: 'hardcoded', stamp: process.env.BEE_POSTAGE as string })
     const app = createApp(
       {
