+++ test/server.spec.ts
@@ -5,7 +5,7 @@ import type { Server } from 'http'
 import { DEFAULT_BEE_DEBUG_API_URL } from '../src/config'
 
 import { bee, getPostageBatch, makeCollectionFromFS } from './utils'
-import { StampsManager } from '../src/stamps'
+import { HardcodedStampsManager } from '../src/stamps'
 import { createHeaderCheckMockServer } from './header-check.mockserver'
 
 const beeApiUrl = process.env.BEE_API_URL || 'http://localhost:1633'
@@ -36,7 +36,7 @@ beforeAll(async () => {
   beeProxy = new Bee(`http://localhost:${port}`)
 
   const stamp = getPostageBatch()
-  const stampManager = new StampsManager()
+  const stampManager = new HardcodedStampsManager()
   await stampManager.start({ mode: 'hardcoded', stamp })
   const appWithStamp = createApp({ beeApiUrl, beeDebugApiUrl: DEFAULT_BEE_DEBUG_API_URL }, stampManager)
   proxyWithStamp = await new Promise((resolve, _reject) => {
